[refactor] 상수 분리, 패키지 이동
