+++ src/main/java/http/request/QueryString.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 import java.util.HashMap;
 
@@ -6,6 +6,8 @@ public class QueryString {
 
     private static final String PARAMETER_DELIMITER = "&";
     private static final String VALUE_DELIMITER = "=";
+    private static final int KEY_INDEX = 0;
+    private static final int VALUE_INDEX = 1;
 
     private HashMap<String, String> param = new HashMap<>();
 
@@ -13,7 +15,7 @@ public class QueryString {
         String[] values = value.split(PARAMETER_DELIMITER);
         for (String s : values) {
             String[] query = s.split(VALUE_DELIMITER);
-            param.put(query[0], query[1]);
+            param.put(query[KEY_INDEX], query[VALUE_INDEX]);
         }
     }
