+++ src/main/java/view/ReturnType.java
@@ -1,7 +1,6 @@
 package view;
 
 public enum ReturnType {
-
     FILE,
     REDIRECT,
     TEMPLATE;
