+++ src/test/java/http/PathTest.java
@@ -1,5 +1,6 @@
 package http;
 
+import http.request.Path;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
