+++ src/test/java/http/QueryStringTest.java
@@ -1,5 +1,6 @@
 package http;
 
+import http.request.QueryString;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
